Moved Managed Stack Tests Into Its Own Class
